Conflicts:
	cube/src/main/java/com/kylinolap/cube/dataGen/FactTableGenerator.java
	metadata/src/main/java/com/kylinolap/metadata/tool/HiveSourceTableMgmt.java
	server/src/main/java/com/kylinolap/rest/controller/TableController.java
	server/src/main/java/com/kylinolap/rest/service/CubeService.java
	storage/src/test/java/com/kylinolap/storage/filter/BitMapFilterEvaluatorTest.java

+++ cube/src/main/java/com/kylinolap/cube/kv/RowValueDecoder.java
@@ -79,13 +79,13 @@ public class RowValueDecoder implements Cloneable {
             Object o = mapredObjs[i];
 
             if (o instanceof LongWritable)
-                o = Long.valueOf(((LongWritable) o).get());
+                o = ((LongWritable) o).get();
             else if (o instanceof IntWritable)
-                o = Integer.valueOf(((IntWritable) o).get());
+                o = ((IntWritable) o).get();
             else if (o instanceof DoubleWritable)
-                o = Double.valueOf(((DoubleWritable) o).get());
+                o = ((DoubleWritable) o).get();
             else if (o instanceof FloatWritable)
-                o = Float.valueOf(((FloatWritable) o).get());
+                o = ((FloatWritable) o).get();
 
             results[i] = o;
         }
